feat(mespapiers): Add translated metadata label to files flexsearchprops
we also move methods from search to helpers
the code change only for `makeFileFlexsearchProps()`
